+++ src/main/java/ru/effective_mobile/tasks/service/impl/TaskServiceImpl.java
@@ -11,9 +11,9 @@ import ru.effective_mobile.tasks.dto.TaskDto;
 import ru.effective_mobile.tasks.model.Task;
 import ru.effective_mobile.tasks.repository.TaskRepository;
 import ru.effective_mobile.tasks.service.TaskService;
+import ru.effective_mobile.tasks.service.UserAuthenticationService;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -23,53 +23,69 @@ public class TaskServiceImpl implements TaskService {
     TaskRepository taskRepository;
     ModelMapper modelMapper;
     ExtendedModelMapper extendedModelMapper;
+    UserAuthenticationService userAuthenticationService;
 
     @Override
-    public Optional<TaskDto> getById(Long id) {
-        return Optional.ofNullable(modelMapper.map(taskRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Task not found")), TaskDto.class));
+    public TaskDto getById(Long id) {
+        return modelMapper.map(taskRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Task not found")), TaskDto.class);
     }
 
     @Override
-    public Optional<List<TaskDto>> getAllByAuthorId(Long id) {
-        Optional<List<Task>> tasksFromRepo = Optional.ofNullable(taskRepository.findAllByAuthorId(id).orElseThrow(() -> new EntityNotFoundException("Tasks not found")));
-        return Optional.ofNullable(extendedModelMapper.mapList(tasksFromRepo.get(), TaskDto.class));
+    public List<TaskDto> getAllByAuthorId(Long id) {
+        List<Task> tasksFromRepo = taskRepository.findAllByAuthorId(id).orElseThrow(() -> new EntityNotFoundException("Tasks not found"));
+        return extendedModelMapper.mapList(tasksFromRepo, TaskDto.class);
     }
 
     @Override
-    public Optional<TaskDto> create(TaskDto taskDto) {
-        return Optional.ofNullable(modelMapper.map(taskRepository.save(modelMapper.map(taskDto, Task.class)),TaskDto.class));
+    public TaskDto create(TaskDto taskDto) {
+        taskDto.setId(null);
+        taskDto.setAuthor(userAuthenticationService.getCurrent());
+        return modelMapper.map(taskRepository.save(modelMapper.map(taskDto, Task.class)), TaskDto.class);
     }
 
     @Override
-    public Optional<TaskDto> edit(TaskDto taskDto) {
-        Optional<TaskDto> taskFromRepo = getById(taskDto.getId());
-        if (taskDto.getDescription() != null) {
-            taskFromRepo.get().setDescription(taskDto.getDescription());
+    public TaskDto edit(TaskDto taskDto) {
+        TaskDto taskFromRepo = getById(taskDto.getId());
+        if (!taskFromRepo.getAuthor().getId().equals(userAuthenticationService.getCurrent().getId())) {
+            throw new SecurityException("You have no access to task with " + taskDto.getId() + " id");
+        } else {
+            if (taskDto.getDescription() != null) {
+                taskFromRepo.setDescription(taskDto.getDescription());
+            }
+            if (taskDto.getStatus() != null) {
+                taskFromRepo.setStatus(taskDto.getStatus());
+            }
+            if (taskDto.getPriority() != null) {
+                taskFromRepo.setPriority(taskDto.getPriority());
+            }
+            if (taskDto.getTitle() != null) {
+                taskFromRepo.setTitle(taskDto.getTitle());
+            }
+            if (taskDto.getPerformer() != null) {
+                taskFromRepo.setPerformer(taskDto.getPerformer());
+            }
+            return modelMapper.map(taskRepository.save(modelMapper.map(taskFromRepo, Task.class)), TaskDto.class);
         }
-        if (taskDto.getStatus() != null) {
-            taskFromRepo.get().setStatus(taskDto.getStatus());
-        }
-        if (taskDto.getPriority() != null) {
-            taskFromRepo.get().setPriority(taskDto.getPriority());
-        }
-        if (taskDto.getTitle() != null) {
-            taskFromRepo.get().setTitle(taskDto.getTitle());
-        }
-        if (taskDto.getPerformerId() != null) {
-            taskFromRepo.get().setPerformerId(taskDto.getPerformerId());
-        }
-        return Optional.ofNullable(modelMapper.map(taskRepository.save(modelMapper.map(taskFromRepo, Task.class)), TaskDto.class));
     }
 
     @Override
-    public Optional<TaskDto> editStatus(TaskDto taskDto) {
-        Optional<TaskDto> taskFromRepo = getById(taskDto.getId());
-        taskFromRepo.get().setStatus(taskDto.getStatus());
-        return Optional.ofNullable(modelMapper.map(taskRepository.save(modelMapper.map(taskFromRepo, Task.class)), TaskDto.class));
+    public TaskDto editStatus(TaskDto taskDto) {
+        TaskDto taskFromRepo = getById(taskDto.getId());
+        if (!taskFromRepo.getPerformer().getId().equals(userAuthenticationService.getCurrent().getId()) || !taskFromRepo.getAuthor().getId().equals(userAuthenticationService.getCurrent().getId())) {
+            throw new SecurityException("You have no access to task with " + taskDto.getId() + " id");
+        } else {
+            taskFromRepo.setStatus(taskDto.getStatus());
+            return modelMapper.map(taskRepository.save(modelMapper.map(taskFromRepo, Task.class)), TaskDto.class);
+        }
     }
 
     @Override
     public void delete(Long id) {
-        taskRepository.deleteById(id);
+        TaskDto taskFromRepo = getById(id);
+        if (!taskFromRepo.getAuthor().getId().equals(userAuthenticationService.getCurrent().getId())) {
+            throw new SecurityException("You have no access to task with " + id + " id");
+        } else {
+            taskRepository.deleteById(id);
+        }
     }
 }
